不要ファイルの削除、それに伴うコードの整理整頓、リファクタリングを行った。 エラー警告が0を確認

+++ Src/Application/GameObject/Textures/Key/KeyBack.cpp
@@ -9,8 +9,8 @@ void KeyBack::Draw()
 	Math::Color color = Math::Color( 0.5f, 0.5f, 0.5f, (float)mi_count / Constant::KEYBACK_COUNT );
 	KdShaderManager::Instance().m_spriteShader.DrawTex(
 		msp_tex.get(),
-		m_pos.x,
-		m_pos.y,
+		(int)m_pos.x,
+		(int)m_pos.y,
 		nullptr,
 		&color )
 		;
